+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 ## 1.0.3
 
-Added persistant storage. When a logged in user closes the app, then reopens, the data
-will now be persistant. Best practice would be to Parse.init, then Parse.currentUser. This
-will return the current user session and allow autologin.
+Added persistent storage. When a logged in user closes the app, then reopens, the data
+will now be persistent. Best practice would be to Parse.init, then Parse.currentUser. This
+will return the current user session and allow auto login. Can also pin data in storage.
 
 ## 1.0.2